Revert "Authentication settings"
