--- API/Entities/ConfigurationModels/JwtConfiguration.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Entities.ConfigurationModels
-{
-    public class JwtConfiguration
-    {
-        public string Section { get; set; } = "JwtSettings";
-        public string? ValidIssuer { get; set; }
-        public string? ValidAudience { get; set; }
-        public string? Expires { get; set; }
-        public string? SecretKey { get; set; }
-        public string? RefreshTokenExpiresDays { get; set; }
-    }
-}
