+++ API/Presentation/Controllers/VideoStatusController.cs
@@ -1,5 +1,4 @@
 ﻿using Entities.Models;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Presentation.ActionFilters;
 using Service.Contracts;
@@ -23,7 +22,6 @@ namespace Presentation.Controllers
             _serviceManager = serviceManager;
         }
 
-        [Authorize]
         [HttpGet]
         public async Task<IActionResult> GetVideoStatuses()
         {
