--- API/Service.Contracts/IAuthenticationService.cs
@@ -1,17 +1,0 @@
-﻿using Shared.DataTransferObjects;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Service.Contracts
-{
-    public interface IAuthenticationService
-    {
-        Task CreateUserAsync(UserForRegistrationDTO userForRegistration);
-        Task<TokenDTO> CreateToken(bool populateExp);
-        Task<TokenDTO> RefreshToken(TokenDTO tokenDTO);
-        Task<bool> ValidateUser(UserForAuthenticationDTO userForAuthentication);
-    }
-}
