+++ API/Service.Contracts/IServiceManager.cs
@@ -10,6 +10,5 @@ namespace Service.Contracts
     {
         IVideoStatusesService VideoStatuses { get; }
         IUserService User { get; }
-        IAuthenticationService Authentication { get; }
     }
 }
