+++ API/Service/ServiceManager.cs
@@ -1,7 +1,5 @@
 ﻿using AutoMapper;
 using Contracts;
-using Entities.ConfigurationModels;
-using Microsoft.Extensions.Options;
 using Repository;
 using Service.Contracts;
 using Service.Helpers;
@@ -17,19 +15,15 @@ namespace Service
     {
         private readonly Lazy<IVideoStatusesService> _videoStatusesService;
         private readonly Lazy<IUserService> _userService;
-        private readonly Lazy<IAuthenticationService> _authenticationService;
 
-        public ServiceManager(IRepositoryManager repositoryManager, IMapper mapper, EntityChecker entityChecker, IOptions<JwtConfiguration> configuration)
+        public ServiceManager(IRepositoryManager repositoryManager, IMapper mapper, EntityChecker entityChecker)
         {
             _videoStatusesService = new Lazy<IVideoStatusesService>(() => new VideoStatusesService(repositoryManager, mapper, entityChecker));
             _userService = new Lazy<IUserService>(() => new UserService(repositoryManager, mapper, entityChecker));
-            _authenticationService = new Lazy<IAuthenticationService>(() => new AuthenticationService(repositoryManager, mapper, configuration, entityChecker));
         }
 
         public IVideoStatusesService VideoStatuses => _videoStatusesService.Value;
 
         public IUserService User => _userService.Value;
-
-        public IAuthenticationService Authentication => _authenticationService.Value;
     }
 }
