--- API/Shared/DataTransferObjects/TokenDTO.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.DataTransferObjects
-{
-    public record TokenDTO(string AccessToken, string RefreshToken);    
-}
