--- API/Shared/DataTransferObjects/UserForAuthenticationDTO.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.DataTransferObjects
-{
-    public record UserForAuthenticationDTO
-    {
-        [Required(ErrorMessage = "Nickname field is required")]
-        public string? Nickname { get; set; }
-
-        [Required(ErrorMessage = "Nickname field is required")]
-        public string? Password { get; set; }
-    }
-}
