+++ API/WebReviews.API/Extensions/ServiceExtensions.cs
@@ -1,12 +1,8 @@
 ﻿using Contracts;
-using Entities.ConfigurationModels;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
 using Repository;
 using Service;
 using Service.Contracts;
-using System.Text;
 
 namespace WebReviews.API.Extensions
 {
@@ -30,35 +26,6 @@ namespace WebReviews.API.Extensions
                     .AllowAnyMethod()
                     .AllowAnyHeader();
                 });
-            });
-
-        public static void ConfigureJwtSetting(this IServiceCollection services, IConfiguration configuration) =>
-            services.Configure<JwtConfiguration>(configuration.GetSection("JwtSettings"));
-
-        public static void ConfigureJwt(this IServiceCollection services, IConfiguration configuration)
-        {
-            var jwtConfiguration = new JwtConfiguration();
-            configuration.Bind(jwtConfiguration.Section, jwtConfiguration);
-
-            services.AddAuthentication(opt =>
-            {
-                opt.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                opt.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            }).AddJwtBearer(options =>
-            {
-                options.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuer = true,
-                    ValidateAudience = true,
-                    ValidateLifetime = true,
-                    ValidateIssuerSigningKey = true,
-                    ClockSkew = TimeSpan.Zero,
-
-                    ValidIssuer = jwtConfiguration.ValidIssuer,
-                    ValidAudience = jwtConfiguration.ValidAudience,
-                    IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtConfiguration.SecretKey))
-                };
-            });
-        }
+            });      
     }
 }
