+++ API/WebReviews.API/appsettings.json
@@ -8,13 +8,5 @@
   "ConnectionStrings": {
     "DefaultConnection": "Host=localhost;Port=5432;Database=WebReviews;Username=postgres;Password=UAZ9233"
   },
-
-  "JwtSettings": {
-    "ValidIssuer": "IRateAPI",
-    "ValidAudience": "IRateHttps",
-    "Expires": 2,
-    "SecretKey": "Secret key which we need to change, mb put in environment",
-    "RefreshTokenExpiresDays": 3
-  },
   "AllowedHosts": "*"
 }
