+++ WebReviews.Tests/Systems/Services/TestUserService.cs
@@ -1,9 +1,7 @@
 ﻿using AutoMapper;
-using Entities.ConfigurationModels;
 using Entities.Exceptions;
 using Entities.Models;
 using FluentAssertions;
-using Microsoft.Extensions.Options;
 using MockQueryable.Moq;
 using Moq;
 using Repository;
@@ -22,7 +20,49 @@ namespace WebReviews.Tests.Systems.Services
 {
     public class TestUserService
     {
-        
+        [Fact]
+        public async Task Get_OnSuccess_Created_User_And_Returned_User()
+        {
+            var fixture = new UserFixture();
+
+            var created = false;
+            var users = fixture.GetRandomData(2).BuildMock().BuildMockDbSet();
+            var userRankGuid = new Guid("3ab56b8e-c3ae-45c3-b9cb-f1a313a61ae5");
+            var userRanks = new List<Userrank>() { new() { UserRankId = userRankGuid, Title = "Бог" } }.BuildMock().BuildMockDbSet();
+
+            var userForCreation = new UserForRegistrationDTO
+            {
+                Nickname = "Oleg",
+                Email = "oleg@mail.ru",
+                Password = "1234",
+                UserRankId = userRankGuid
+            };
+
+            var mockAutoMapper = new MapperConfiguration(cfg =>
+            {
+                cfg.AddProfile(new MappingProfiles());
+            });
+
+            var autoMapper = mockAutoMapper.CreateMapper();
+
+            var mockContext = new Mock<WebReviewsContext>();
+            mockContext.Setup(x => x.Set<User>()).Returns(users.Object);
+            mockContext.Setup(x => x.Set<User>().Add(It.IsAny<User>())).Callback(() =>
+            {
+                created = true;
+            });
+            mockContext.Setup(x => x.Set<Userrank>()).Returns(userRanks.Object);
+
+            var repositoryManager = new RepositoryManager(mockContext.Object);
+            var entityChecker = new EntityChecker(repositoryManager);
+
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
+
+            await serviceManager.User.CreateUserAsync(userForCreation);
+
+            created.Should().BeTrue();
+            mockContext.Verify(x => x.Set<User>().Add(It.IsAny<User>()), Times.Once());
+        }
 
         [Fact]
         public async Task Get_OnSuccess_Deleted_User()
@@ -49,9 +89,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             await serviceManager.User.DeleteUserAsync(guid, trackChanges: true);
 
@@ -81,9 +120,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var listOfUsers = await serviceManager.User.GetUsersAsync(trackChanges: false);
 
@@ -111,9 +149,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var user = await serviceManager.User.GetUserByIdAsync(guid, trackChanges: false);
 
@@ -149,9 +186,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var updateUser = await serviceManager.User.UpdateUserAsync(guid, userForUpdate, trackChanges: true);
 
@@ -186,9 +222,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             await serviceManager.Invoking(async x => await x.User.UpdateUserAsync(guid, userForUpdate, trackChanges: true))
                 .Should().ThrowAsync<NotFoundException>();
@@ -224,9 +259,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var updateUser = await serviceManager.User.UpdateUserAsync(guid, userForUpdate, trackChanges: true);
