+++ WebReviews.Tests/Systems/Services/TestVideoStatusesService.cs
@@ -1,9 +1,7 @@
 ﻿using AutoMapper;
-using Entities.ConfigurationModels;
 using Entities.Exceptions;
 using Entities.Models;
 using FluentAssertions;
-using Microsoft.Extensions.Options;
 using MockQueryable.Moq;
 using Moq;
 using Repository;
@@ -49,9 +47,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             await serviceManager.VideoStatuses.CreateVideoStatusAsync(referenceDTO);
 
@@ -82,9 +79,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var entity = await serviceManager.VideoStatuses.GetVideoStatusByIdAsync(guid, trackChanges: true);
 
@@ -114,9 +110,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var entities = await serviceManager.VideoStatuses.GetVideoStatusesAsync(trackChanges: true);
 
@@ -147,9 +142,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             var entitiy = await serviceManager.VideoStatuses.UpdateVideoStatus(guid, referenceForManipulation, trackChanges: true);
 
@@ -184,9 +178,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             await serviceManager.VideoStatuses.DeleteVideoStatusAsync(guid, trackChanges:true);
 
@@ -217,9 +210,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             //var entitiy = await serviceManager.VideoStatuses.UpdateVideoStatus(guid, referenceForManipulation, trackChanges: true);
 
@@ -249,9 +241,8 @@ namespace WebReviews.Tests.Systems.Services
 
             var repositoryManager = new RepositoryManager(mockContext.Object);
             var entityChecker = new EntityChecker(repositoryManager);
-            var options = Options.Create(new JwtConfiguration());
 
-            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker, options);
+            var serviceManager = new ServiceManager(repositoryManager, autoMapper, entityChecker);
 
             //var entitiy = await serviceManager.VideoStatuses.UpdateVideoStatus(guid, referenceForManipulation, trackChanges: true);
